drivers: hid: G940 FF structure & autocenter fixes
Disable default auto-centering of the Logitech G940, which otherwise
makes the stick difficult to use. Also when autocenter is requested, obey
the magnitude.

The auto-centering effect will only unlock (or become the requested
level) when an application starts using the joystick and the hand sensor
on the stick is covered. That appears to be a hardware limitation.

To make the above and future work on spring/damper support simpler and
more type-safe, add a struct of the HID report. Many thanks to fred41 from
forums.x-plane.org for useful information towards that.

It seems byte 0 does not need to be a "command byte" of 0x51; it can be
anything and is the least-significant byte of the constant force.

Signed-off-by: Chris Boyle <chris@boyle.name>